app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
APScheduler==3.11.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
lxml==5.4.0
multidict==6.4.4
numpy==2.3.0
pandas==2.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.1
